% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC_MS_Performance.R
\name{Calculate_run_order_statistics}
\alias{Calculate_run_order_statistics}
\title{test2.}
\usage{
Calculate_run_order_statistics(QQC)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
