% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeAndTransform.R
\name{CollapseToProtein}
\alias{CollapseToProtein}
\title{Add two numbers.}
\usage{
CollapseToProtein(QQC, opt, norm = "ref")
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
