% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduce.R
\name{ComputeUMAP}
\alias{ComputeUMAP}
\title{Computes UMAP}
\usage{
ComputeUMAP(QQC)
}
\arguments{
\item{QQC}{a QuantQC object}
}
\value{
A \code{QQC object} with the reductions slot containing a UMAP data.frame in list
}
\description{
This function takes a QQC object and computes UMAP on the imputed protein level data with the UMAP function
from the Seurat package
}
\examples{
ComputeUMAP(TestSamples)
}
