% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_statistics.R
\name{Count_peptides_per_cell}
\alias{Count_peptides_per_cell}
\title{Add two numbers.}
\usage{
Count_peptides_per_cell(sc.data, cellenONE_meta, good_cells = NULL)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
