% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeAndTransform.R
\name{DIA_carrier_norm}
\alias{DIA_carrier_norm}
\title{Add two numbers.}
\usage{
DIA_carrier_norm(Raw_data_lim.d, carrier_CH, plex_used)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
