% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduce.R
\name{FeaturePCA}
\alias{FeaturePCA}
\title{Computes PCA}
\usage{
FeaturePCA(QQC, prot = NA, imputed = T)
}
\arguments{
\item{QQC}{a QuantQC object}
}
\value{
A \code{QQC object} with the reductions slot containing a PCA data.frame in list
}
\description{
This function takes a QQC object and computes PCA on the unimputed protein level data taking the eigen
values of a correlation matrix computed off pairwise observations.
}
\examples{
ComputePCA(TestSamples)
}
