% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantQC_functions.R
\name{Gen_QQC_report_from_list}
\alias{Gen_QQC_report_from_list}
\title{test1.}
\usage{
Gen_QQC_report_from_list(data_path, linker_path, isolation1, isolation2 = NULL)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
