% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantQC_functions.R
\name{inSet_completness}
\alias{inSet_completness}
\title{Add two numbers.}
\usage{
inSet_completness(Raw_data, cellenOne_meta)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
