% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellenONE_metadata_map.R
\name{link_manual_Raw}
\alias{link_manual_Raw}
\title{Link Manual Annotations to raw data.}
\usage{
link_manual_Raw(QQC)
}
\arguments{
\item{QQC}{A QuantQC object.}
}
\value{
\code{QQC} object with linked sample meta data
}
\description{
This function takes a QuantQC (QQC) object as an input and return a QQC object
that has sample meta data linked in the meta.data slot. Its only to be used if
cellenONE files are not available
}
\examples{
link_manual_Raw(TestSamples)
}
