% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantQC_functions.R
\name{Normalize_reference_vector_log}
\alias{Normalize_reference_vector_log}
\title{Add two numbers.}
\usage{
Normalize_reference_vector_log(dat)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
