% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduce.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Plots PCA}
\usage{
PlotPCA(QQC, by = "Condition")
}
\arguments{
\item{QQC}{a QuantQC object}

\item{by}{a string that specifies how to color code PCA}
}
\value{
A \code{QQC object} with the reductions slot containing a PCA data.frame in list
}
\description{
This function takes a QQC object and plots PC1 and PC2 as a scatter plot. It can be colored by different
options like sort condition, or various potential batch effect sorces
}
\examples{
ComputePCA(TestSamples)
}
