% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantQC_functions.R
\name{PlotProtAndPep}
\alias{PlotProtAndPep}
\title{Add two numbers.}
\usage{
PlotProtAndPep(nPOP_obj)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
