% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellenONE_metadata_map.R
\name{PlotSlideLayout_celltype}
\alias{PlotSlideLayout_celltype}
\title{Plot cellenONE droplets by cell types}
\usage{
PlotSlideLayout_celltype(QQC)
}
\arguments{
\item{QQC}{A QQC object}
}
\value{
A plot
}
\description{
This function takes plots the positions of the single cell droplets over glass slide colored
by the type of sample sorted. Wells where sample is placed are written in text in center of drop array.
}
\examples{
PlotSlideLayout_celltype(TestSamples)
}
