% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellenONE_metadata_map.R
\name{PlotSlideLayout_label}
\alias{PlotSlideLayout_label}
\title{Plot cellenONE droplets by labels}
\usage{
PlotSlideLayout_label(QQC)
}
\arguments{
\item{QQC}{A QQC object}
}
\value{
A plot
}
\description{
This function takes plots the positions of the single cell droplets over glass slide colored
by the label used.
}
\examples{
PlotSlideLayout_celltype(TestSamples)
}
