% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeAndTransform.R
\name{remove_sets_carrier_lessthan_SCs}
\alias{remove_sets_carrier_lessthan_SCs}
\title{Add two numbers.}
\usage{
remove_sets_carrier_lessthan_SCs(df, files)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add_numbers(2, 3)
}
